correction of GruntFile to point on telephony
Correction of events to remove log of event contents when in production
In config parameter a log level at root level of logs in config : config.logs.level has been added for set the level of logs. Note: The value in file section will erase it, if the file logging is enabled.
Default log level is now info
refactor of telephony service start to follow promise system
In RESTService, core, Channels, contacts remove logs of result content when requesting data
in XMPPService remove logs of result content when requesting data and when event is received (maybe it is not complete, may need more refactoring)
Refactor xmpp ping to log the rainbow sdk version. (Nice for debug files from other party)