Add ChannelsService::updateChannelVisibility method to update a channel visibility (closed or company)
Add ChannelsService::updateChannelVisibilityToPublic method to update a channel visibility to company (visible for users in that company)
Add ChannelsService::updateChannelVisibilityToClosed method to update a channel visibility to closed (not visible by users)
Add ChannelsService::updateChannelTopic method to update the description of the channel to update (max-length=255)
Add ChannelsService::updateChannelName method to update a channel name
Update AdminService::createUserInCompany to add the roles parameter when creating a user.
Reduce log for the XMPP ping.