add method getStatusForUser in Bubble class to get the status of a user in the bubble
update the "leaveBubble" method to unsubscribe form a bubble if not alreadyu unsubscribed the connected user. Else delete it from bubble.
Fix file url in XMPPService::sendChatExistingFSMessage and XMPPService::sendChatExistingFSMessageToBubble methods
Add method BubblesService::getStatusForConnectedUserInBubble to Get the status of the connected user in a bubble
Fix ImsService::getMessageFromBubbleById method