Add event evt_internal_bubbleavatarchanged raised when a bubble is updated with a new avatar
Add avatar property in Bubble class. This is the URL to download the avatar . Note that it can spot on an empty file if no avatar has been setted.
Add BubblesService::promoteContactToModerator Promote a contact to moderator in a bubble
Add BubblesService::demoteContactFromModerator Demote a contact to user in a bubble
Add BubblesService::getAvatarFromBubble Get A Blob object with data about the avatar picture of the bubble.
Add BubblesService::updateAvatarForBubble  Update the bubble avatar (from given URL). The image will be automaticalle resized.
Add BubblesService::deleteAvatarFromBubble Delete the bubble avatar
Add private methods in BubblesService : randomString, setAvatarBubble, deleteAvatarBubble, resizeImage, getBinaryData
Add BubblesService::updateCustomDataForBubble Update the customData of the bubble
Add BubblesService::deleteCustomDataForBubble Delete the customData of the bubble
Add BubblesService::updateDescriptionForBubble Update the description of the bubble. (it is the topic on server side, and result event)
Add BubblesService::openConversationForBubble Open a conversation to a bubble