Add Utils::loadConfigFromIniFile Utils::saveConfigFromIniFile to manage a config.ini file in $USERDATA$/Rainbow/RainbowNodeSdkDir directory.
Update 'xmppRessourceName' Option to be store in config.ini file to keep the same ressource name when a new SDK is constructed. So a user can only be connected one times on one computer. Note that the ressource name will be the same if few users has an instance of SDK.